requirements.txt
cassandra-driver==3.25.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.0.3
geomet==0.2.1.post1
idna==3.3
memory-profiler==0.60.0
numpy==1.22.3
psutil==5.9.1
Pympler==1.0.1
requests==2.28.1
six==1.16.0
tabulate==0.8.9
urllib3==1.26.11
